+++ src/stock-trading-vue/static/StockTradingApp.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+import LoginView from './views/LoginView.vue'
+
+</script>
+
+<template>
+    <LoginView />
+</template>