+++ src/stock-trading-vue/static/views/LoginView.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import Login from '../components/user/Login.vue';
+</script>
+
+<template>
+    <div class="d-flex align-items-center py-4">
+        <main class="form-signin w-100 m-auto">
+            <Login />
+        </main>
+    </div>
+</template>
+
+<style src="../assets/sign-in.css"></style>